ignore election browser when creating visitor, move testPass out of testFolder
