+++ cypress/integration/Dashboard/AI-reports.test.js
@@ -1,4 +1,4 @@
-import testPass from '../testPass'
+import testPass from '../../testPass'
 
 describe('AI Reports', () => {
   it('should load', () => {
