+++ cypress/integration/Dashboard/articles.test.js
@@ -1,4 +1,4 @@
-import testPass from '../testPass'
+import testPass from '../../testPass'
 
 describe('Articles', () => {
   it('should load', () => {
