+++ cypress/integration/Dashboard/logout.test.js
@@ -1,4 +1,4 @@
-import testPass from '../testPass'
+import testPass from '../../testPass'
 
 describe('Log out', () => {
   it('should load', () => {
