+++ cypress/integration/Dashboard/suggestions.test.js
@@ -1,4 +1,4 @@
-import testPass from '../testPass'
+import testPass from '../../testPass'
 
 describe('Suggestions', () => {
   it('should load', () => {
